made POST request that takes email and min ACT
