+++ app.py
@@ -1,6 +1,6 @@
 from flask import Flask
 from flask_restful import Resource, reqparse, Api
-from collegeDataEmail import cdgmain, send_email
+from collegeDataEmail import college_data_email, send_email
 
 app = Flask(__name__)
 api = Api(app)
@@ -15,6 +15,12 @@ db.create_all()
 
 class Movies_List(Resource):
     parser = reqparse.RequestParser()
+    
+    """
+    parser.add_argument('min_ACT', type=int, required=False, help='Minimum acceptable ACT Score')
+    parser.add_argument('gender_distribution', type=int, required=False, help='Desired gender distribution')
+    parser.add_argument('email', type=str, required=True, help='email address of the recipient')
+    """
     parser.add_argument('director', type=str, required=False, help='Director of the movie')
     parser.add_argument('genre', type=str, required=False, help='Genre of the movie')
     parser.add_argument('collection', type=int, required=True, help='Gross collection of the movie')
@@ -50,15 +56,42 @@ class Movies_List(Resource):
             item.delete_()
             return {'Message': '{} has been deleted from records'.format(movie)}
         return {'Message': '{} is already not on the list'.format()}
+
+
+class Email_Data(Resource):
+    parser = reqparse.RequestParser()
     
+    """
+    parser.add_argument('min_ACT', type=int, required=False, help='Minimum acceptable ACT Score')
+    parser.add_argument('gender_distribution', type=int, required=False, help='Desired gender distribution')
+    parser.add_argument('email', type=str, required=True, help='email address of the recipient')
+    """
+    #parser.add_argument('director', type=str, required=False, help='Director of the movie')
+    parser.add_argument('min_ACT', type=int, required=False, help='Genre of the movie')
+    parser.add_argument('recipient_email', type=str, required=True, help='Gross collection of the movie')
+
+     
+    def get(self):
+        return {'class': 'Email_Data'}
+
+
+    def post(self):
+        args = Email_Data.parser.parse_args()
+        
+        college_data_email(args['recipient_email'], "personalized.csv", args['min_ACT'])
+
+        return {'output': "sent!"}
+
+
 class All_Movies(Resource):
     def get(self):
-        filename = "27PageAll.csv"
-        cdgmain(filename)
-        send_email("jackdavidweber@gmail.com", filename)
-        return {'output': "sent!"}
+        filename = "schoolNames.csv"
+        #cdgmain(filename)
+        #send_email("jackdavidweber@gmail.com", filename)
+        return {'output': "Nothing for now"}
     
 api.add_resource(All_Movies, '/')
+api.add_resource(Email_Data, '/send-data')
 api.add_resource(Movies_List, '/<string:movie>')
 
 if __name__=='__main__':
